Moved feature flag setup up code down to the NuGet.Jobs.Common. It is still optional.
