+++ src/NuGet.Jobs.Common/FeatureFlags/FeatureFlagRefresher.cs
@@ -6,9 +6,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using NuGet.Jobs.Configuration;
 using NuGet.Services.FeatureFlags;
 
-namespace NuGet.Jobs.Validation
+namespace NuGet.Jobs
 {
     public class FeatureFlagRefresher : IFeatureFlagRefresher
     {
