+++ src/NuGet.Jobs.Common/LoggerDiagnosticsSource.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using NuGet.Services.Logging;
 using NuGetGallery.Diagnostics;
 
-namespace NuGet.Jobs.Validation
+namespace NuGet.Jobs
 {
     public class LoggerDiagnosticsSource : IDiagnosticsSource
     {
