+++ src/NuGet.Jobs.Common/NuGet.Jobs.Common.csproj
@@ -1,14 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="..\..\SdkProjects.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>net472;netstandard2.0</TargetFrameworks>
+    <TargetFrameworks>net472;netstandard2.1</TargetFrameworks>
     <RootNamespace>NuGet.Jobs</RootNamespace>
     <Description>Common infrastructure for running the NuGetGallery back-end jobs.</Description>
   </PropertyGroup>
   
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
     <Compile Remove="Configuration\MessageServiceConfiguration.cs" />
   </ItemGroup>
 
@@ -33,6 +33,12 @@
     </PackageReference>
     <PackageReference Include="NuGet.Services.Sql">
       <Version>$(ServerCommonPackageVersion)</Version>
+    </PackageReference>
+    <PackageReference Include="NuGet.Services.FeatureFlags">
+      <Version>$(ServerCommonPackageVersion)</Version>
+    </PackageReference>
+    <PackageReference Include="NuGetGallery.Core">
+      <Version>$(NuGetGalleryPackageVersion)</Version>
     </PackageReference>
   </ItemGroup>
