+++ src/NuGet.Services.Validation.Orchestrator/Job.cs
@@ -227,13 +227,9 @@ namespace NuGet.Services.Validation.Orchestrator
             });
             services.AddTransient<ISharedAccessSignatureService, SharedAccessSignatureService>();
 
-            /// See <see cref="SubscriptionProcessorJob{T}.ConfigureDefaultJobServices(IServiceCollection, IConfigurationRoot)"/>
-            /// for reasoning on why this is registered here.
-            services.AddSingleton<IFeatureFlagRefresher, FeatureFlagRefresher>();
-
             ConfigureFileServices(services, configurationRoot);
             ConfigureOrchestratorSymbolTypes(services);
-            ValidationJobBase.ConfigureFeatureFlagServices(services, configurationRoot);
+            JsonConfigurationJob.ConfigureFeatureFlagServices(services, configurationRoot);
         }
 
         protected override void ConfigureAutofacServices(ContainerBuilder containerBuilder, IConfigurationRoot configurationRoot)
@@ -305,7 +301,7 @@ namespace NuGet.Services.Validation.Orchestrator
                     throw new NotImplementedException($"Unknown type: {validatingType}");
             }
 
-            ValidationJobBase.ConfigureFeatureFlagAutofacServices(containerBuilder);
+            JsonConfigurationJob.ConfigureFeatureFlagAutofacServices(containerBuilder);
             ConfigureLeaseService(containerBuilder);
         }
