+++ tests/NuGet.Jobs.Common.Tests/FeatureFlags/FeatureFlagRefresherFacts.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
@@ -6,12 +9,12 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Moq;
-using NuGet.Jobs.Validation;
+using NuGet.Jobs.Configuration;
 using NuGet.Services.FeatureFlags;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Validation.Common.Job.Tests.FeatureFlags
+namespace NuGet.Jobs.Common.Tests.FeatureFlags
 {
     public class FeatureFlagRefresherFacts
     {
